Fix conftest.
