+++ tests/test_extract.py
@@ -4,7 +4,6 @@ import json
 import os
 import pandas as pd
 
-from tests.utils import datadir
 import dgpost.utils
