+++ tests/test_load.py
@@ -1,10 +1,7 @@
 import pytest
-import yadg.core
 import json
 import os
-import pandas as pd
 
-from tests.utils import datadir
 import dgpost.utils
