+++ tests/test_parse.py
@@ -1,10 +1,6 @@
 import pytest
-import yadg.core
-import json
 import os
-import pandas as pd
 
-from tests.utils import datadir
 import dgpost.utils
