+++ tests/utils.py
@@ -1,27 +1,7 @@
 import pytest
 import os
 import json
-from distutils import dir_util
 import yadg.core
-
-
-@pytest.fixture
-def datadir(tmpdir, request):
-    """
-    from: https://stackoverflow.com/a/29631801
-    Fixture responsible for searching a folder with the same name of test
-    module and, if available, moving all contents to a temporary directory so
-    tests can use them freely.
-    """
-    filename = request.module.__file__
-    test_dir, _ = os.path.splitext(filename)
-    if os.path.isdir(test_dir):
-        dir_util.copy_tree(test_dir, str(tmpdir))
-    base_dir, _ = os.path.split(test_dir)
-    common_dir = os.path.join(base_dir, "common")
-    if os.path.isdir(common_dir):
-        dir_util.copy_tree(common_dir, str(tmpdir))
-    return tmpdir
 
 
 def datagram_from_input(input, parser, datadir):
